+++ CHANGELOG.md
@@ -11,11 +11,20 @@ Changelog
 
 _Note: Gaps between patch versions are faulty, broken or test releases._
 
-## v3.xx.x (2021-06-01)
+## v3.xx.x (2021-06-04)
 
 #### :rocket: New Feature
 
+* Added possibility to listen to events from a function emitter `core/async`
 * Added option to change size of `restricted cache`
+
+#### :bug: Bug Fix
+
+* Fixed `Date.create` with dates without leading zeroes
+
+#### :house: Internal
+
+* Updated dependencies: `typescript@4.3.2`
 
 ## v3.49.2 (2021-05-27)