+++ src/core/async/CHANGELOG.md
@@ -9,6 +9,12 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.xx.x (2021-06-04)
+
+#### :rocket: New Feature
+
+* Added possibility to listen to events from a function emitter
+
 ## v3.42.1 (2021-04-12)
 
 #### :house: Internal